feat[task being added to projects]
Created a logic in the addTaskToProject() that will
select a project and add task to it in accordance
with the form details/submit.

Upon task submission addTaskToProject() will be called
and add task to the project for the user.

We are adding actual  object (newTask) from
submitTaskButtonFunction() hence return{ newTask }
needed.